# docs/swagger.yaml
basePath: /
definitions:
  api.GenerateKeyRequest:
    properties:
      email:
        example: user@example.com
        type: string
      label:
        example: Production
        type: string
      scopes:
        description: Optional scopes for this key (read|write|export|admin). Defaults to read,write,export.
        items:
          type: string
        type: array
    type: object
  api.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  api.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  api.ResetConfirmRequest:
    properties:
      email:
        example: user@example.com
        type: string
      new_password:
        example: newsecurepassword456
        type: string
      token:
        example: a1b2c3
        type: string
    type: object
  api.ResetRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
...
  store.APIKey:
    properties:
      created_at:
        type: integer
      expires_at:
        type: integer
      id:
        description: Stable key identifier (sha256 hex of the raw key).
        type: string
      is_revoked:
        type: boolean
      key:
        description: Raw API key. Returned only on create/rotate; not persisted in plaintext.
        type: string
      key_last4:
        description: Redacted display helper for UI/logging.
        type: string
      key_prefix:
        description: Redacted display helper for UI/logging.
        type: string
      label:
        type: string
      last_used_at:
        type: integer
      scopes:
        description: Optional scopes for this key (read|write|export|admin).
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@velarix.dev
    name: API Support
    url: http://velarix.dev/support
  description: Decision integrity API for AI-assisted internal approvals. Velarix records facts, tracks causal dependencies, and invalidates stale reasoning before downstream actions are taken.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://velarix.dev/terms/
  title: Velarix API
  version: 0.1.0
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT for console access. No authentication
        required.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        '200':
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        '401':
          description: invalid credentials
          schema:
            type: string
      summary: Login user
      tags:
      - Auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with Argon2id password hashing. No authentication
        required.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        '201':
          description: user created
          schema:
            additionalProperties:
              type: string
            type: object
        '400':
          description: invalid request
          schema:
            type: string
        '500':
          description: internal error
          schema:
            type: string
      summary: Register a new user
      tags:
      - Auth
  /v1/auth/reset-confirm:
    post:
      consumes:
      - application/json
      description: Update password using a reset token provided via console logs.
        No authentication required.
      parameters:
      - description: Reset details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ResetConfirmRequest'
      produces:
      - application/json
      responses:
        '200':
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        '401':
          description: invalid token
          schema:
            type: string
      summary: Confirm password reset
      tags:
      - Auth
  /v1/auth/reset-request:
    post:
      consumes:
      - application/json
      description: Generate a reset token and log it to the server console. No authentication
        required.
      parameters:
      - description: User email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ResetRequest'
      produces:
      - application/json
      responses:
        '200':
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset
      tags:
      - Auth
...
  /v1/keys:
    get:
      consumes:
      - application/json
      description: Retrieve all API keys associated with the user. Admin role required.
      parameters: []
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.APIKey'
            type: array
        '401':
          description: unauthorized
          schema:
            type: string
        '403':
          description: forbidden
          schema:
            type: string
        '404':
          description: user not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: List API keys
      tags:
      - Auth
...
  /v1/keys/{key}:
    delete:
      consumes:
      - application/json
      description: Mark an API key as revoked. Revoked keys return 401.
      parameters:
      - description: Key identifier (sha256 hex). For backwards compatibility, legacy raw keys may still be accepted.
        example: '"b7c3f5...sha256hex..."'
        in: path
        name: key
        required: true
        type: string
      - description: Optional idempotency key for safe retries.
        in: header
        name: Idempotency-Key
        required: false
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        '401':
          description: unauthorized
          schema:
            type: string
        '404':
          description: key not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: Revoke an API key
      tags:
      - Auth
  /v1/keys/{key}/rotate:
    post:
      consumes:
      - application/json
      description: Revokes the old key and generates a new one with the same label.
        Requires admin role.
      parameters:
      - description: Old key identifier (sha256 hex). For backwards compatibility, legacy raw keys may still be accepted.
        example: '"b7c3f5...sha256hex..."'
        in: path
        name: key
        required: true
        type: string
      - description: Optional idempotency key for safe retries.
        in: header
        name: Idempotency-Key
        required: false
        type: string
      produces:
      - application/json
      responses:
        '201':
          description: key rotated
          schema:
            $ref: '#/definitions/store.APIKey'
        '401':
          description: unauthorized
          schema:
            type: string
        '403':
          description: forbidden
          schema:
            type: string
        '404':
          description: key not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: Rotate an API key
      tags:
      - Auth
  /v1/keys/generate:
    post:
      consumes:
      - application/json
      description: Create a new labeled API key for the user. Supports optional scoped keys (read|write|export|admin).
      parameters:
      - description: Key details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GenerateKeyRequest'
      - description: Optional idempotency key for safe retries.
        in: header
        name: Idempotency-Key
        required: false
        type: string
      produces:
      - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/store.APIKey'
        '400':
          description: invalid request
          schema:
            type: string
        '403':
          description: forbidden
          schema:
            type: string
        '404':
          description: user not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: Generate a new API key
      tags:
      - Auth
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: '2.0'